requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
category-encoders==2.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.1.1
entrypoints==0.4
executing==2.0.1
Faker==24.9.0
fastapi==0.110.1
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
ipython==8.23.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
orjson==3.10.0
packaging==24.0
pandas==2.1.4
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
plotly-resampler==0.10.0
pmdarima==2.0.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pycaret==3.3.0
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7.1
pyod==1.1.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
schemdraw==0.15
scikit-base==0.7.5
scikit-learn==1.4.1.post1
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
sktime==0.28.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
streamlit==1.33.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tbats==1.1.3
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
tsdownsample==0.1.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
validators==0.28.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xgboost==2.0.3
xxhash==3.4.1
yellowbrick==1.5
zipp==3.18.1
